+++ Src/GUI/Plugins/PlotterData.py
@@ -8,7 +8,8 @@ import netCDF4
 import h5py
 
 from MDANSE.Core.Decorators import compatibleabstractproperty
-from MDANSE.IO.IOUtils import HDFFileVariable, netcdf_find_numeric_variables
+from MDANSE.IO.NetCDF import find_numeric_variables as netcdf_find_numeric_variables
+from MDANSE.IO.HDF5 import find_numeric_variables as hdf_find_numeric_variables
 
 
 class _IPlotterData:
@@ -64,44 +65,12 @@ class HDFPlotterData(_IPlotterData):
 
         self._variables = collections.OrderedDict()
 
-        HDFPlotterData.find_numeric_variables(self._variables, self._file)
+        hdf_find_numeric_variables(self._variables, self._file)
 
     @property
     def variables(self):
         return self._variables
 
-    @staticmethod
-    def find_numeric_variables(var_dict, group):
-        """This method retrieves all the numeric variables stored in the NetCDF file.
-
-        This is a recursive method.
-        """
-
-        for var_key, var in group.items():
-
-            if isinstance(var,h5py.Group):
-                HDFPlotterData.find_numeric_variables(var_dict,var)
-            else:
-
-                if var.parent.name == '/':
-                    path = '/{}'.format(var_key)
-                else:
-                    path = '{}/{}'.format(var.parent.name, var_key)
-
-                # Non numeric variables are not supported by the plotter
-                if not numpy.issubdtype(var.dtype, numpy.number):
-                    continue
-
-                # Variables with dimension higher than 3 are not supported by the plotter
-                if var.ndim > 3:
-                    continue
-
-                comp = 1
-                while var_key in var_dict:
-                    var_key = '{:s}_{:d}'.format(var_key, comp)
-                    comp += 1
-
-                var_dict[var_key] = (path,HDFFileVariable(var))
 
 PLOTTER_DATA_TYPES = {
     '.nc': NetCDFPlotterData,
